refactor: moved context menu primitives up a folder
