+++ view/shared/ui/src/Menu/CheckboxItem.tsx
