+++ view/shared/ui/src/Menu/Content.tsx
