+++ view/shared/ui/src/Menu/index.ts
