+++ view/shared/ui/src/Menu/types.ts
